Extract syscall_open() handler
