+++ src/core/sys/syscall.c
@@ -144,32 +144,3 @@ void syscall_read(registers_t* registers)
   desc->offset += bytes_read;
   registers->rdx = bytes_read;
 }
-
-void syscall_open(registers_t* registers)
-{
-  errno = 0;
-
-  const char* pathname = (const char*)registers->rbx;
-  uint32_t flags = registers->rcx;
-
-  inode_t inode = vfs_namei(pathname);
-
-  if (inode == 0) {
-    registers->rdx = -1;
-    errno = ENOENT;
-    return;
-  }
-
-  int fd = create_file_descriptor(inode, flags);
-
-  if (fd == -1) {
-    CORE_SYS_DEBUG("%s", "too many files open");
-    registers->rdx = -1;
-    errno = ENFILE;
-    return;
-  }
-
-  registers->rdx = fd;
-
-  CORE_SYS_DEBUG("open fd=%d inode=%p flags=%d", registers->rdx, inode, flags);
-}
